herbarium_records: move HerbariumRecordForm to Views/
Components::HerbariumRecordForm -> Views::Controllers::HerbariumRecords::Form
Per #4331. Also updates a HerbariumRecordForm reference in
autocompleter_field_test.rb. All component, controller, integration,
and observation_show system tests pass.
